Restructue gmac module
